add api  update inscription

+++ src/main/java/tn/essatin/erp/rest/DiplomeEtudiantRest.java
@@ -10,6 +10,7 @@ import tn.essatin.erp.dao.DiplomeEtudiantDao;
 import tn.essatin.erp.dao.EtudiantsDao;
 import tn.essatin.erp.model.DiplomeEtudiant;
 import tn.essatin.erp.model.Etudiants;
+import tn.essatin.erp.payload.request.AjouterDiplomeEtudiantRequest;
 import tn.essatin.erp.payload.request.DiplomeEtudiantByIdEtudiantRequest;
 import tn.essatin.erp.payload.request.DiplomeEtudiantByIdRequest;
 import tn.essatin.erp.payload.request.ModifierDiplomeEtudiantRequest;
@@ -23,12 +24,9 @@ import java.util.List;
 @RequestMapping("/api/diplomeetudiant/")
 public class DiplomeEtudiantRest {
 
-    final
-    DiplomeEtudiantDao diplomeEtudiantDao;
-    final
-    EtudiantsDao etudiantsDao;
-    final
-    DiplomeDao diplomeDao;
+    final DiplomeEtudiantDao diplomeEtudiantDao;
+    final EtudiantsDao etudiantsDao;
+    final DiplomeDao diplomeDao;
 
     @Autowired
     public DiplomeEtudiantRest(DiplomeEtudiantDao diplomeEtudiantDao, EtudiantsDao etudiantsDao, DiplomeDao diplomeDao) {
@@ -39,12 +37,15 @@ public class DiplomeEtudiantRest {
 
     @PostMapping("/getbyid")
     public ResponseEntity<?> getById(@Valid @RequestBody DiplomeEtudiantByIdRequest diplomeEtudiantByIdRequest) {
-        return new ResponseEntity<>(diplomeEtudiantDao.findById(diplomeEtudiantByIdRequest.getIdDiplomeEtudiant()).get(), HttpStatus.OK);
+        if (diplomeEtudiantDao.findById(diplomeEtudiantByIdRequest.getIdDiplomeEtudiant()).isPresent())
+            return new ResponseEntity<>(diplomeEtudiantDao.findById(diplomeEtudiantByIdRequest.getIdDiplomeEtudiant()).get(), HttpStatus.OK);
+        else
+            return ResponseEntity.ok(new MessageResponse("diplome introuvable!", 204));
     }
 
     @PostMapping("/getbyidetudiant")
     public ResponseEntity<?> getByIdEtudiant(@Valid @RequestBody DiplomeEtudiantByIdEtudiantRequest diplomeEtudiantByIdEtudiantRequest) {
-        Etudiants e = new Etudiants();
+        Etudiants e;
         if (etudiantsDao.findById(diplomeEtudiantByIdEtudiantRequest.getIdEtudiant()).isPresent()) {
             e = etudiantsDao.findById(diplomeEtudiantByIdEtudiantRequest.getIdEtudiant()).get();
             List<DiplomeEtudiant> l = diplomeEtudiantDao.findByIdEtudiant(e);
@@ -58,22 +59,51 @@ public class DiplomeEtudiantRest {
 
     @PostMapping("/supprimerbyid")
     public ResponseEntity<?> supprimerById(@Valid @RequestBody DiplomeEtudiantByIdRequest diplomeEtudiantByIdRequest) {
-        diplomeEtudiantDao.delete(diplomeEtudiantDao.findById(diplomeEtudiantByIdRequest.getIdDiplomeEtudiant()).get());
-        return ResponseEntity.ok(new MessageResponse("Diplome Etudiant Supprimer avec succée!!"));
+        if (diplomeEtudiantDao.findById(diplomeEtudiantByIdRequest.getIdDiplomeEtudiant()).isPresent()) {
+            diplomeEtudiantDao.delete(diplomeEtudiantDao.findById(diplomeEtudiantByIdRequest.getIdDiplomeEtudiant()).get());
+            return ResponseEntity.ok(new MessageResponse("Diplome Etudiant Supprimer avec succée!!"));
+        } else
+            return new ResponseEntity<>(new MessageResponse("Diplome introuvable", 403), HttpStatus.FORBIDDEN);
     }
+
 
     @PostMapping("/modifer")
     public ResponseEntity<?> modiferById(@Valid @RequestBody ModifierDiplomeEtudiantRequest modifierDiplomeEtudiantRequest) {
-        DiplomeEtudiant de = diplomeEtudiantDao.findById(modifierDiplomeEtudiantRequest.getIdDiplome()).get();
-        de.setIdDiplome(diplomeDao.findById(modifierDiplomeEtudiantRequest.getIdDiplome()).get());
-        de.setAnnee(modifierDiplomeEtudiantRequest.getAnnee());
-        de.setEtablissement(modifierDiplomeEtudiantRequest.getEtablissement());
-        de.setNiveau(modifierDiplomeEtudiantRequest.getNiveau());
-        de.setSpecialite(modifierDiplomeEtudiantRequest.getSpecialite());
-        de.setStatus(modifierDiplomeEtudiantRequest.getStatus());
-        diplomeEtudiantDao.save(de);
+        if (
+                diplomeEtudiantDao.findById(modifierDiplomeEtudiantRequest.getIdDiplome()).isPresent()
+                        && diplomeDao.findById(modifierDiplomeEtudiantRequest.getIdDiplome()).isPresent()
+        ) {
+            DiplomeEtudiant de = diplomeEtudiantDao.findById(modifierDiplomeEtudiantRequest.getIdDiplome()).get();
+            de.setIdDiplome(diplomeDao.findById(modifierDiplomeEtudiantRequest.getIdDiplome()).get());
+            de.setAnnee(modifierDiplomeEtudiantRequest.getAnnee());
+            de.setEtablissement(modifierDiplomeEtudiantRequest.getEtablissement());
+            de.setNiveau(modifierDiplomeEtudiantRequest.getNiveau());
+            de.setSpecialite(modifierDiplomeEtudiantRequest.getSpecialite());
+            de.setStatus(modifierDiplomeEtudiantRequest.getStatus());
+            diplomeEtudiantDao.save(de);
+            return ResponseEntity.ok(new MessageResponse("Diplome Etudiant modifier avec succée!!"));
+        } else
+            return new ResponseEntity<>(new MessageResponse("Ressources introuvable", 403), HttpStatus.FORBIDDEN);
+    }
 
-        return ResponseEntity.ok(new MessageResponse("Diplome Etudiant modifier avec succée!!"));
+    @PostMapping("/ajouter")
+    public ResponseEntity<?> Ajouter(@Valid @RequestBody AjouterDiplomeEtudiantRequest ajouterDiplomeEtudiantRequest) {
+        if (
+                etudiantsDao.findById(ajouterDiplomeEtudiantRequest.getIdEtudiant()).isPresent()
+                        && diplomeDao.findById(ajouterDiplomeEtudiantRequest.getIdDiplome()).isPresent()
+        ) {
+            DiplomeEtudiant de = new DiplomeEtudiant();
+            de.setIdDiplome(diplomeDao.findById(ajouterDiplomeEtudiantRequest.getIdDiplome()).get());
+            de.setIdEtudiant(etudiantsDao.findById(ajouterDiplomeEtudiantRequest.getIdEtudiant()).get());
+            de.setAnnee(ajouterDiplomeEtudiantRequest.getAnnee());
+            de.setSpecialite(ajouterDiplomeEtudiantRequest.getSpecialite());
+            de.setNiveau(ajouterDiplomeEtudiantRequest.getNiveau());
+            de.setStatus(ajouterDiplomeEtudiantRequest.getStatus());
+            de.setEtablissement(ajouterDiplomeEtudiantRequest.getEtablissement());
+            diplomeEtudiantDao.save(de);
+            return ResponseEntity.ok(new MessageResponse("Diplome Etudiant Ajouter avec succée!!", 200));
+        } else
+            return new ResponseEntity<>(new MessageResponse("Ressources introuvable", 403), HttpStatus.FORBIDDEN);
     }
 
 }
